Rework the landing page
OpenStack Foundation renamed itself in 2020, and
the bylaws are changing in 2025 as a part of the
Foundation's move under the Linux Foundation.

The board no longer maintains its old wiki page,
it's moved its business to a website.

And let's prefix "OpenStack" to governing body names
so we can distinguish between other top level projects
under the OIF.

Change-Id: Idc0bc3778ecfc7749e4acf40b99b35503bd47580
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>